fix handling of response timeouts
previously, because the response function didn't support cooperative
cancellation, even after a timeout we would still hang waiting for the response
to arrive.

withThrowingTimeout() now takes a callback to be invoked on timeout which we
can use to resume the response continuation with a timed out error.

We also move the timeout handling into the core sendCommandRrq() function.