OcaConnectionBroker: add device expiration grace period before removing
add a grace period before actually removing a device from the broker, in case
it gets re-added. temporarily disable reconnection during this period in case
the device actually has disappeared, but do not disconnect from it until the
grace period has expired and we have confirmed the device has not reappeared.